requirements.txt
streamlit
torch
transformers
sentence-transformers
pdf2image
Pillow
